requirements.txt
category_encoders==2.5.1.post0
fastapi==0.88.0
gradio==3.12.0
numpy==1.21.5
pandas==1.4.3
protobuf==4.21.10
pymongo==4.2.0
python-dotenv==0.21.0
schedule==1.1.0
scikit_learn==1.1.1
scipy==1.7.3
google-api-python-client==2.61.0
google-cloud-storage==2.5.0
dnspython==2.2.1